refactor : Package명 변경
Oracle package 명 정책에 따라 이름을 AutoLotto에서 Lotto로 변경
