+++ src/main/java/lotto/Lotto.java
@@ -1,4 +1,4 @@
-package autolotto;
+package lotto;
 
 import java.util.ArrayList;
 import java.util.Collections;
