+++ src/main/java/lotto/OutputView.java
@@ -1,4 +1,4 @@
-package autolotto;
+package lotto;
 
 import java.util.Collections;
 import java.util.List;
