+++ src/test/java/lotto/LottoTest.java
@@ -1,4 +1,4 @@
-package autolotto;
+package lotto;
 
 import org.junit.jupiter.api.Test;
