+++ src/test/java/lotto/LottoTicketTest.java
@@ -1,4 +1,4 @@
-package autolotto;
+package lotto;
 
 import org.junit.jupiter.api.Test;
